# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Offloadr API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /projects/{id}/mark-review:
    post:
      operationId: markProjectReview
      tags: [projects]
      summary: Mark project as needing review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project marked for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/archive:
    post:
      operationId: archiveProject
      tags: [projects]
      summary: Archive a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/missing-files:
    get:
      operationId: getMissingFiles
      tags: [projects]
      summary: Get missing file checklist for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Missing files checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissingFilesResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        projectName:
          type: string
        episodeTitle:
          type: string
          nullable: true
        clientName:
          type: string
          nullable: true
        recordingDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, uploading, review_needed, ready_for_editor, delivered, archived]
        description:
          type: string
          nullable: true
        editorNotes:
          type: string
          nullable: true
        expectedCameraCount:
          type: integer
          nullable: true
        expectedAudioSetup:
          type: string
          nullable: true
        fileCount:
          type: integer
        totalSize:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - projectName
        - status
        - fileCount
        - totalSize
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        projectName:
          type: string
        episodeTitle:
          type: string
          nullable: true
        clientName:
          type: string
          nullable: true
        recordingDate:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        editorNotes:
          type: string
          nullable: true
        expectedCameraCount:
          type: integer
          nullable: true
        expectedAudioSetup:
          type: string
          nullable: true
      required:
        - projectName
    UpdateProjectBody:
      type: object
      properties:
        projectName:
          type: string
        episodeTitle:
          type: string
          nullable: true
        clientName:
          type: string
          nullable: true
        recordingDate:
          type: string
          nullable: true
        status:
          type: string
        description:
          type: string
          nullable: true
        editorNotes:
          type: string
          nullable: true
        expectedCameraCount:
          type: integer
          nullable: true
        expectedAudioSetup:
          type: string
          nullable: true
    MissingFilesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              present:
                type: boolean
              required:
                type: boolean
            required:
              - label
              - present
              - required
        allPresent:
          type: boolean
      required:
        - items
        - allPresent
...